surfaceflinger: Fix hwcomposer command errors
When using the hwcomposer API, there are two paths that the command may
take. One path results in a direct synchronous IPC message, whereas the
other path will buffer the commands and dispatch later just before
present.

+++ services/surfaceflinger/DisplayHardware/ComposerHal.cpp
@@ -255,6 +255,11 @@ Error Composer::createVirtualDisplay(uint32_t width, uint32_t height,
 
 Error Composer::destroyVirtualDisplay(Display display)
 {
+    Error error = execute();
+    if (error != Error::NONE) {
+        return error;
+    }
+
     auto ret = mClient->destroyVirtualDisplay(display);
     return unwrapRet(ret);
 }
@@ -284,6 +289,11 @@ Error Composer::createLayer(Display display, Layer* outLayer)
 
 Error Composer::destroyLayer(Display display, Layer layer)
 {
+    Error error = execute();
+    if (error != Error::NONE) {
+        return error;
+    }
+
     auto ret = mClient->destroyLayer(display, layer);
     return unwrapRet(ret);
 }
